Fix WinBoard game-list title
Number of displayed games and WDL stats were always appended to the
existing title, leading to an ever-growing title in the game-list
window. Now the old actual title is ignored, and the starts is appended
to the original title "Game List". (In the future we could use the
filename here?)

In addition, the title was not updated when re-opening the game-list
window, while the stats could change if filtering was active when the
window was closed. (In that case the filter is cleared on re-opening, so
the number of displayed games changes.) Now a title-update (after
recalculating the stats) is done every time the game-list window is
opened.