Define /firstUCI and /secondUCI as synonyms for /fUCI, /sUCI
Because the %s expansion of WB options in the /fcp and /scp lines only
works for options that contain 'first' or 'second', and /fUCI thus did
not work, PSWBTM had to rely on the cumbersome /firstIsUCI=true to
install UCI engines. (The latter option should be deprecated really;
the UCI-ness of the engines is a volatile option, so it makes no sense
to supply a way to set them both to false or true. One only needs to be
able to set it to the non-default state.)