Tricked by the grossnes
Some WB menu items were referenced by a hard number. As items had been
added over the years, the wrong menu items were grayed. In particular
the Options -> Adjudication menu was grayed out, in -ncp mode, where
this was intended for the Sounds dialog. This accidentally proved
useful, until we started to apply adjudictions also to the Human moves.
Then it was suddenly undesirable, and a small mystery why it happened at
all, when I tried to change it. The numbers are now corrected to the new
locatons of the items in the menus.