Revert splitting of UserMoveEvent
This splitting into UserMoveTest and FinishMove was no longer needed,
now that the new mouse driver no longer calls UserMoveTest to figure out
if a move is a promotion, and XBoard and WinBoard are unified in this
respect. It was counter-productive, because UserMoveTest was not able to
change the fromX, fromY (call-by-value) parameters for passing them to
FinishMove, so it had to defer the conversion of drop moves to internal
format to the latter, while it did not belong there. (Because
DropMenuEvent already delivers the moves in internal format.) This
whole kludgy construction was therefore abandoned, by going back
to the original 4.2.7 calling scheme, where UserMoveEvent completely
pre-processes the move, and when it is valid, calls FinishMove directly.