Make board-size overrule options volatile
The options -boardWidth, -boardHeight and -holdingSize were legacies
from the day choosing a variant would not automatically set a board
size. As a result they are seldomly used anymore. The first real
application is mini-Shogi (on 5x5 board), and it turned out very
annoying if the 5x5 settings persisted to the next session, where you
almost certainly want to play another variant.