Add set of shogi pixmaps for XBoard
The -size 64 pieces from gShogi (and originally XShogi) are made
suitable for loading by XBoard, through redefining the transparent
background color as either light or dark square color. These bitmaps are
not included in the XBoard build, and perhaps should not even be
included in the source distribution, but made available in a separate
distribution?