Make starting new variation dependent on shift key
The shift key is read during mouse events, and posted in a global
variable, which can then be used by the back-end. The decision to start
a new variation (pushing the game tail rather than clipping it off) when
entring a move in Edit Game or Analyze mode is made dependent on this.
The move-typein also records the shift-key state.