Implement board textures in XBoard
The -darkBackTextureFile and -liteBackTextureFile options now wrk in
XBoard too, for loading a pixmap that will be used to cut out the dark
and light square backgrounds. The cutting algorithm is such that the
pixmap image is completely restored if the size was that of the total
board (e.g. a drawing of a Xiangqi board).