Let -oneClickMove also work in EditGame mode
There is really no reason why people that lke one-clcik moves enough to
switch them on would not want to use them too when editing a game.
This solves a nast inconsistency, as before it used to work in -ncp mode
on the _first_ move, as the mode is then BeginningOfGame, and not yet
EditGame, and we do want it to work for BeginningOfGame, as this is
also the game mode when you start playing as white.