Animate piece explosions in drag-drop moves and in XB
This required some code restructuring: the decision if an explosion is
needed is now taken in a new routine Explode() in the backend. This then
calls the front-end driver, and it returns the info if there was an
explosion or not, so the caller (AnimateMove() or the mouse driver) can
take action to repair the damage to the board caused by the blast wave.
A front-end driver for XBoard is provided as well, so that both
click-click (and replay, which is the same) and drag-drop atomic
captures are animated in XB as well as WB.
Explosions on rejected moves are suppressed.