Augment moves of some fairy pieces
To get better disambiguation in variants that have to be played with
legality testing off, the Elephant is given also one-step moves (making
it a so-called Modern Elephant) in -variant fairy and great. The Dragon
King (Crowned Rook), which had no moves defined for it outside Shogi at
all, has been given Rook moves plus an orthogonal 2-jump, so that it
can get SAN moves when used as Dababba or Short Rook. (Illegal moves
currently aways force long algebraic notation...)