Fix spurious promotions with legality testing off
The recent promotion patch had broken playing without legality testing
completely: without legality testing one always gets the piece that was
asked for, but a NULLCHAR should not be taken as a request for a piece!
But it was, and this resulted in any moved piece turning into a white
Pawn! For true promotions the piece was always black, which was also not
very useful.

Also revert to the use of '+' as internal Shogi promoChar, because
pre-processing turns out not to be possible when reading a game file.
In stead parse any trailing '+' on moves as promoChar. Downside: '++'
can no longer be recognized as checkmate symbol.