Keep last PV while clearing engine-output display
When ponder is on, the display is cleared immediately when the last line
of thinking output appears by the appearing ponder output. So you would
never get the chance to see it. This line is now re-written after
clearing, so it appears at the bottom of the output for the next search.