Fix bug in parsing illegal Pawn captures
When a pawn move of type exf5 was illegal, the parser made an attempt to
interpret it as e.p. capture of f5. This, however, messed up the move,
so that it was now permanently mistaken for e5xf6. Even in cases with
legality testing off, where you would want the original move to be
accepted despite the fact that it was illegal.