Implement setup (engine-GUI) command
This allows the frst engine (when legality testing is off) to overrule
the default opening position, in reply to the variant command. Syntax is
"setup (PIECE_TO_CHAR_TABLE) FEN_STRING"
where the FEN string is interpreted in the context of the given
pieceToCharTable (which is optional, and only delimeted by parethesis if
it is present).
Sendng the position to the second engine is delayed until the first
engine has responded to the ping after the new-game sequence, to make
sure it gets the chance to alter the position before it is sent to the
second engine. This feature is useful for playing variants that WinBoard
does not even know the opening positon for, and can be used even if they
are shuffle variants.