Let Engine #2 Settings start second engine
If the second engine is not yet running, it would be impossible to use
the settings dialog for it, as the features have not been received and
the dialog will be empty. So the user could not adjust the engine
parameters before starting a match. Using this menu now automatically
starts the engine, and delays the appearence of the popup untill the
features are received, using a similar mechanism as in
TwoMachinesEventIfReady().