Fix one-click moving on up-click
(First) up-clicks were not always ignored: in Edit-Game mode with
one-click moving enabled this could lead to a capture being made on
the down-click, and the recapture on the correspoding up-click.
This has now been corrected, and only the down-click of a first click
can lead to selecting a piece or a one-click move.