Fix spurious scores in comments
The comment popup is also used to display engine PV info, but when the
contents of the window is written back when we select a variation,
it was treated as a real comment. This is fixed now by checing in
ReplaceComment() if the comment starts with the PV info, and if so,
ignore the latter. This solves the problem, but I guess the real bug is
that selecting a variation tries to write the comment even if it did not
change.