Bugfix XBoard menu translation
The gettext macro N_() was applied to the menu definition tables, but
this would break any translation, as these items are both used as menu
texts as well as reference for enabling and checkmarking. The routines
that create the menus already apply _() dynamically to these items when
they are used as menu texts.