Separate menu text from menu name in XBoard
The text string in the menu-definition tables was used both for the menu
text, as well as for referencing the items from the code (for enabing or
checkmarking). Now these can be defined separately, so that altering the
menu texts is possible without breaing any references to the item.
(Moving them to another menu would still break XBoard, though!)
Currently the texts and names are still the same.