Adapt some XBoard menu texts
The "Show" prefix is removed from the "View" menu items, as this seems
implied. An ellipsis is added to the Engine #N Settings items, which
could not be there before, because periods in a menu name would break
any reference to it. For some other names the ellipsis was deleted.
(These were not refered to, but who knows what we will do in the future.)