Update texi file for new menu organization
The descriptions are shuffled around and organized into sections
corresponding to the new main menus. Where necessary the name of the
items was changed to their new names. The items removed from the File
menu (the next/previous/same stuff) are nowmoved to the section "other
shortcut keys", as the key bindings for some of them are still active.