Improve drawing of highlight arrow
A diagonal arrow sometimes had a notch in the shaft because of
independent rounding of the starting and ending point coordinates.
Furthermore the width of the shaft was unnecessary variable(as a
function of direction) because half the width was rounded and then used
twice. Now rounding of end-points is always done towards a fixed grid
of square centers, which keeps diagonals purely diagonal, and one side
of the shaft is drawn relative to the other side, after the
position of the latter has been rounded, so any rounding error affects
the shaft width only once.