Allow line-straddling result comments
Some GUIs (in particular Arena) do not have the decency to keep the
result comment on a single line. In this case the parser rules would not
recognize it, and treat it as a normal comment, plus an undetailed PGN
result. The rules are now fixed to allow linefeeds in the comment, and
between comment and PGN result. To correctly process the comment further
(e.g. in game-list lines) the linefeeds are replaced by spaces in the
back-end routines that do use the result comment.