Peel PV out of comment
Arena stores engine PVs not in variation format, but within both braces
and parentheses, before the score/depth info. This caused XBoard to
overlook the score/depth info, (frustrating, e.g., the eval graph),
while the PV would not be accessible for walking it as a variation.
The routine that extracts the score/depth from a comment now has
been made more clever, to recognize a leadin PV in the comment, and then
place it in front of the comment as variation (by moving the opening
brace to behind it). For some reason it does not work if the PV and
score/depth were separated by a newline, but this happens quite rarely.