Configure EGTB cache size in adapterCommand
The hash size, nr of CPUs and EGTB path are transferred through the WB
protocol interface commands memory, cores and egtpath. But there is no
WB command for detailing the EGTB cache size. So the value of the
command-line option -defaultCacheSizeEGTB has to be sent to Polyglot
through its startup command to make it known to UCI engines.