Provide menus for editing WinBoard engine and server lists
The Edit Tags dialog in the WB front-end is adapted to act as a general
edit for string variables, by giving it an extra parameter to indicate
where to store the edited result. Menu items are added to invoke this
dialog for editing the /first, /secondChessProgramNames and /icsNames
options. This solves the problem that people would have to edit an ini
file that they can no longer locate (because the %APPDATA% folder has
different names and locations in the various Windows versions, and is
often a hidden folder on top of that).
There is a slight memory leak here, as the memory for the previous
value of the altered string option is not freed.