Subject WB context menus to translation
The innerloopof the menu translation has been made into a separate
routine. This now uses GetMenuItemInfo() in stead of ModifyMenu() to
better preserve menu breaks. Menu #10 is used for context menus (from
the routine that pops them up), and for this number the orignal texts
are not saved in menuText. (This is not needed, because the menu is
created everytime anew from the resource data.)