Disable some accelerators in JAWS version
Single-key accelerators that could result in irreversible data loss are
undesirable, because the risk of accidentally pushing them is too large.
So the non-Alt duplicates of next/previous game/position are disabled,
as are Home and End for Revert and Truncate Game. JAWS users will have
to rely on the menus for those functions.