Add variant Spartan Chess
In the WB menu it takes the place of the defunct Twilight Chess item.
The setup routine needed patching for setting up Lances in stead of
Pawns for black, the check test had to be adapted to handle duple check,
Alfil and Dragon moves were fine tuned in this variant to make the exact
moves, and the promotion popup was adapted to show King and interpret
Rook and Bishop as Dragon and Alfil for black. The Alfil now always
moves as Modern Elephant except in Shatranj and Courier.