Fishy fix
The move that delivers checkmate in two-machines mode is not animated
when adjudication is on. There then also is no highlighting of this move
Redrawing the board and calling ShowMove fixes it. Which is rather
strange, because ShowMove was already called in Adjudicate (when it
adjudicates), before GameEnds. For reasons that are not understood this
call to ShowMove() (which calls AnimateMove()) is not effective.