Fix time in PGN info
Saving of the thinking time in the PGN was broken by the patch that
removed the race condition in the clock switching, because that moved
the incrementing of forwardMostMove to behind the code that used fmm to
index pvInfoList. As a result the time was stored in the slot for the
opponent's previous move. The value is corrected now.