Fix logo repainting
The logos were not repainted when the main window was moved out of the
display area, and then back into view. The reason for this is
unknown, as the PaintProc calls HDCDrawPosition, which was calling
DrawLogoOnDC. But the logos were not drawn, (or on an expose event
sometimes half drawn!) despite the fact that the rest of the board was.
The drawing of logos now has been decoupled from drawing the board,
and has been moved to a separate routine DisplayLogos() (in analogy to
DisplayBothClocks()). This is cleaner anyway, as the logos usually do
not change when the board does. So apart from in PaintProc, it has only
to be called when the mode changes (ModeHighlight()) or on swap clocks.
This seems to fix the problem.