Fix legality testing for promotions
Turns out WB has never tested for the legality of the chosen promotion
piece, and would allow promotion to King innorml chess even with
legality testing on (when entered through the keyboard)! Now strict
testing is done, outlawing Pawns and Kings (or Unicorns in
Knightmate). Only in suicide / giveaway Kings are stillallowed, and in
spartan a black King if you do not have two yet.(And promotion
toLnce,representing Hoplite Pawn,is forbidden there.)