Fix type-in of drop moves
Drop moves entered through the move type-in already come in the (piece,
DROP_RANK, x, y) format, rather than having a holdings from square. This
caused UserMoveEvent to consider them illegal when testing if the
piecewas of the side to move, and when this was fixed, to recognize
Knight drops as coming from between board and holdings (and thus
illegal).