Fix WB Sound Options greyout and remove some grossness
The sound and ICS Options menu items in WB were accessed 'by position',
with fixed defined positions (always messing up any reorganization of
the menus when you forgot to redefine the position constants in
winboard.h). And not only that, but for completely mysterious reasons,
the greying of the sound item greyed out the game-list options (above
it) in stead when it was defined as 9, and the ICS options (below it)
when it was 10 !!!. Now they are addressed 'by command' (and the
position definitions are removed), which fixes things.