Move clock-click code to back-end
Some duplicated (or should-be-duplicated) code for handling clicks on
the clocks from the XBoard and WinBoard front-ends was merged and
relocated to the back-end, so it can be shared. Noticeable was that
XBoard allowed turn change in ICS examine mode by clock clicking,
while WinBoard did not. WB, on the other hand, allowed clock adjustment.