Greyout unavailable variants in New Variant dialog
The WB radio-button code is refactored to be driven from a list (which
must be kept in phase with the variant definitions in common.h!). It now
also checks if the first engine supports the variant, and greys out the
item if it isn't. XB now also makes this check. Slight re-ordering of
10x8 variants there, to have the more common on top.
If no radio button is selected, keep old variant even on OK.