Update texi file
The highlight-with-arrow and one-click-move menu items are described.
The highlightMoveWithArrow and variations option are described.
Load Next /Previous Position descriptions are moved to File menu.
The description for entering variations is updated (using Shift key).
The remark that editing holdings is impossible is removed.
Spartan Chess is added to the list of supported variants.