Clean up fishy patch
None of the ShowMove() calls in Adjudicate() was working, because when
this code was taken out of HandleMachineMove() it was taken outside the
scope of the local fromX, fromY etc. there, so that it started using the
global variables from the mouse driver. As these usually contain -1,
this suppressed move animation and highlighting. All these calls have
now been removed, and ShowMove() is called after a positive return (from
HandleMachineMove(), so within the correct scope).
This improves the previous fix, which led to some flicker, because the
erroneous ShowMove() calls did draw the new board (instantly), wich in
the fishy patch was then immediately overwritten by the old position, to
start animation. This drawing of the old position is now no longer
needed either.