Fix Alfil bug
The test for Alfil (suppressing single steps on the Elephant) contained
a wrong logical operator, so that it could never be active. It also
failed to include Xiangqi as variant with traditional Alfil. This led to
one-step moves of the Elephant in these variants, and false check or
unreliable mate detections.

Fix move of Xiangqi Elephant