Fix showing of user move after adjudication
The correction of the fishy patch (removing ShowMove() from Adjudicate())
had broken the adjudication of user moves (that actually did work before
the fishy patch, because there the gobal fromX/Y and toX/Y do contain
the correct values for the move to be shown). An entered move that led
to adjudication (like a checkmate) now was not displayed at all. Adding
an extra ShowMove() after (successful) return of Adudicate() solves it.