Fix some resource leaks
Some files were opened not closed in XBoard, when loading XIM images
(which I doubt anyone ever does) but also when copying the clipboard
(which has the potential of being done very often). There was also some
code that would not work if conditional code which is permanently
disabled would be enabled.