Fix flipBlack option XBoard with board texture
In XBoard -flipBlack is implemented by swapping white and black pieces,
because the Shogi bitmaps from XShogi are made that way. But the mask
used to cut out the whole in the board texture to fit the piece was not
swapped, both when drawing static pieces, as well as during animation.