Fix two bugs in reading position diagram
When reading a game without FEN tag, a position diagram is used to set
the initial position. This calls CharToPiece() with a '.' as argument,
because that is the position-diagram way to indicate empty squares.
altered CharToPiece() over time this would be recognized as the first
undefined piece or, more recently, the first piece without nickName,
which is a white Pawn. Now it returns EmptySquare again in that case.
A second problem was that for position diagrams in braces the brace was
not in the list of ignored characters, and would cause the reading to
get out of phase, so that the side to move was not recognized.