Fix parsing bug of FRC castling
The code that was supposed to intercept castling when no rights existed
in FRC was not adapted to the new encoding for this by NoRights, rather
than -1, and furthermore returned 0 (= EndOfFile) rather than
ImpossibleMove. Not sure if the former would manifest itself in any
way, but the latter seems bad.