Implement Grand Chess
Grand Chess requires a lot of exceptional handling similar to Superchess
and Great Shatran, because it also allows only promotion to a captured
piece. In addition it shares the enlarged promotion zone with Makruk and
Shogi, as well as the fact that promotion can be deferred. The 10-rank
board furthermore required a better generalization of double-pushes and
e.p. captures. On such a board double puhes are now allowed from 2nd and
3rd rank, e.p. capture from 5th and 4th.