Add board dialog XBoard
The GenericPopUp() is used to implement a dialog for board options. The
board dialog needs to trigger redraw and redefinition of the
colors, through an OK callback. Color-pickers are used for the 6 basic
color settings.
  To allow changing board settings during the session (i.e. from a menu
dialog), the routines to allocate resources have to free the resources
for the previous settings first. This patch take care of XtCreateGCs(),
preventing re-creation of resources that are unchangeable (such as the
grid color), and calling XtReleastGC() on GCs that can be changed.
To prevent resource leaks the old pixmaps are deleted before making new
ones on second or later calls of CreateXPMPieces().