Add file-browser option to just return name
By passing mode "f" the file browser dialog will not open the selected
file at all, so it can be used to retreive the name for the benefit of a
file-browser button to a -file option. A mode "p" will limit the
selection to directories only, to browse for a path.